condaenv.evpokour.requirements.txt
sqlalchemy
alembic
pymysql
cryptography
fastapi
uvicorn
watchfiles
aiofiles
python-multipart
starlette
selenium
websocket-client
playwright
aiohttp
ffmpeg-python
pydantic
pydantic-core
pyinstaller
pillow
imagehash